+++ plugins/inputs/graylog/graylog.go
@@ -76,53 +76,6 @@ func (c *RealHTTPClient) HTTPClient() *http.Client {
 	return c.client
 }
 
-var sampleConfig = `
-  ## API endpoint, currently supported API:
-  ##
-  ##   - multiple  (e.g. http://<host>:9000/api/system/metrics/multiple)
-  ##   - namespace (e.g. http://<host>:9000/api/system/metrics/namespace/{namespace})
-  ##
-  ## For namespace endpoint, the metrics array will be ignored for that call.
-  ## Endpoint can contain namespace and multiple type calls.
-  ##
-  ## Please check http://[graylog-server-ip]:9000/api/api-browser for full list
-  ## of endpoints
-  servers = [
-    "http://[graylog-server-ip]:9000/api/system/metrics/multiple",
-  ]
-
-  ## Set timeout (default 5 seconds)
-  # timeout = "5s"
-
-  ## Metrics list
-  ## List of metrics can be found on Graylog webservice documentation.
-  ## Or by hitting the web service api at:
-  ##   http://[graylog-host]:9000/api/system/metrics
-  metrics = [
-    "jvm.cl.loaded",
-    "jvm.memory.pools.Metaspace.committed"
-  ]
-
-  ## Username and password
-  username = ""
-  password = ""
-
-  ## Optional TLS Config
-  # tls_ca = "/etc/telegraf/ca.pem"
-  # tls_cert = "/etc/telegraf/cert.pem"
-  # tls_key = "/etc/telegraf/key.pem"
-  ## Use TLS but skip chain & host verification
-  # insecure_skip_verify = false
-`
-
-func (h *GrayLog) SampleConfig() string {
-	return sampleConfig
-}
-
-func (h *GrayLog) Description() string {
-	return "Read flattened metrics from one or more GrayLog HTTP endpoints"
-}
-
 // Gathers data for all servers.
 func (h *GrayLog) Gather(acc telegraf.Accumulator) error {
 	var wg sync.WaitGroup