+++ plugins/inputs/lustre2/lustre2.go
@@ -32,23 +32,6 @@ type Lustre2 struct {
 	allFields map[tags]map[string]interface{}
 }
 
-var sampleConfig = `
-  ## An array of /proc globs to search for Lustre stats
-  ## If not specified, the default will work on Lustre 2.5.x
-  ##
-  # ost_procfiles = [
-  #   "/proc/fs/lustre/obdfilter/*/stats",
-  #   "/proc/fs/lustre/osd-ldiskfs/*/stats",
-  #   "/proc/fs/lustre/obdfilter/*/job_stats",
-  #   "/proc/fs/lustre/obdfilter/*/exports/*/stats",
-  # ]
-  # mds_procfiles = [
-  #   "/proc/fs/lustre/mdt/*/md_stats",
-  #   "/proc/fs/lustre/mdt/*/job_stats",
-  #   "/proc/fs/lustre/mdt/*/exports/*/stats",
-  # ]
-`
-
 /* The wanted fields would be a []string if not for the
 lines that start with read_bytes/write_bytes and contain
    both the byte count and the function call count
@@ -446,16 +429,6 @@ func (l *Lustre2) GetLustreProcStats(fileglob string, wantedFields []*mapping) e
 		}
 	}
 	return nil
-}
-
-// SampleConfig returns sample configuration message
-func (l *Lustre2) SampleConfig() string {
-	return sampleConfig
-}
-
-// Description returns description of Lustre2 plugin
-func (l *Lustre2) Description() string {
-	return "Read metrics from local Lustre service on OST, MDS"
 }
 
 // Gather reads stats from all lustre targets