+++ plugins/inputs/mcrouter/mcrouter.go
@@ -29,15 +29,6 @@ const (
 	typeFloat statType = iota
 )
 
-var sampleConfig = `
-  ## An array of address to gather stats about. Specify an ip or hostname
-  ## with port. ie tcp://localhost:11211, tcp://10.0.0.1:11211, etc.
-	servers = ["tcp://localhost:11211", "unix:///var/run/mcrouter.sock"]
-
-	## Timeout for metric collections from all servers.  Minimum timeout is "1s".
-  # timeout = "5s"
-`
-
 var defaultTimeout = 5 * time.Second
 
 var defaultServerURL = url.URL{
@@ -111,16 +102,6 @@ var sendMetrics = map[string]statType{
 	"cmd_other_count":                            typeInt,
 	"cmd_delete_out_all":                         typeInt,
 	"cmd_lease_set_out_all":                      typeInt,
-}
-
-// SampleConfig returns sample configuration message
-func (m *Mcrouter) SampleConfig() string {
-	return sampleConfig
-}
-
-// Description returns description of Mcrouter plugin
-func (m *Mcrouter) Description() string {
-	return "Read metrics from one or many mcrouter servers"
 }
 
 // Gather reads stats from all configured servers accumulates stats