+++ plugins/inputs/mongodb/mongostat.go
@@ -149,6 +149,8 @@ type WiredTiger struct {
 	Transaction TransactionStats       `bson:"transaction"`
 	Concurrent  ConcurrentTransactions `bson:"concurrentTransactions"`
 	Cache       CacheStats             `bson:"cache"`
+	Connection  WTConnectionStats      `bson:"connection"`
+	DataHandle  DataHandleStats        `bson:"data-handle"`
 }
 
 // ShardStats stores information from shardConnPoolStats.
@@ -246,6 +248,16 @@ type StorageEngine struct {
 type TransactionStats struct {
 	TransCheckpointsTotalTimeMsecs int64 `bson:"transaction checkpoint total time (msecs)"`
 	TransCheckpoints               int64 `bson:"transaction checkpoints"`
+}
+
+// WTConnectionStats stores statistices on wiredTiger connections
+type WTConnectionStats struct {
+	FilesCurrentlyOpen int64 `bson:"files currently open"`
+}
+
+// DataHandleStats stores statistics for wiredTiger data-handles
+type DataHandleStats struct {
+	DataHandlesCurrentlyActive int64 `bson:"connection data handles currently active"`
 }
 
 // ReplStatus stores data related to replica sets.
@@ -743,6 +755,12 @@ type StatLine struct {
 	ModifiedPagesEvicted      int64
 	UnmodifiedPagesEvicted    int64
 
+	// Connection statistics (wiredtiger only)
+	FilesCurrentlyOpen int64
+
+	// Data handles statistics (wiredtiger only)
+	DataHandlesCurrentlyActive int64
+
 	// Replicated Opcounter fields
 	InsertR, InsertRCnt                      int64
 	QueryR, QueryRCnt                        int64
@@ -1142,6 +1160,10 @@ func NewStatLine(oldMongo, newMongo MongoStatus, key string, all bool, sampleSec
 		returnVal.UnmodifiedPagesEvicted = newStat.WiredTiger.Cache.UnmodifiedPagesEvicted
 
 		returnVal.FlushesTotalTime = newStat.WiredTiger.Transaction.TransCheckpointsTotalTimeMsecs * int64(time.Millisecond)
+
+		returnVal.FilesCurrentlyOpen = newStat.WiredTiger.Connection.FilesCurrentlyOpen
+
+		returnVal.DataHandlesCurrentlyActive = newStat.WiredTiger.DataHandle.DataHandlesCurrentlyActive
 	}
 	if newStat.WiredTiger != nil && oldStat.WiredTiger != nil {
 		returnVal.Flushes, returnVal.FlushesCnt = diff(newStat.WiredTiger.Transaction.TransCheckpoints, oldStat.WiredTiger.Transaction.TransCheckpoints, sampleSecs)