+++ plugins/inputs/tcp_listener/tcp_listener.go
@@ -58,20 +58,6 @@ var dropwarn = "tcp_listener message queue full. " +
 var malformedwarn = "tcp_listener has received %d malformed packets" +
 	" thus far."
 
-const sampleConfig = `
-  # DEPRECATED: the TCP listener plugin has been deprecated in favor of the
-  # socket_listener plugin
-  # see https://github.com/influxdata/telegraf/tree/master/plugins/inputs/socket_listener
-`
-
-func (t *TCPListener) SampleConfig() string {
-	return sampleConfig
-}
-
-func (t *TCPListener) Description() string {
-	return "Generic TCP listener"
-}
-
 // All the work is done in the Start() function, so this is just a dummy
 // function.
 func (t *TCPListener) Gather(_ telegraf.Accumulator) error {