+++ plugins/inputs/teamspeak/teamspeak.go
@@ -3,9 +3,10 @@ package teamspeak
 import (
 	"github.com/multiplay/go-ts3"
 
+	"strconv"
+
 	"github.com/influxdata/telegraf"
 	"github.com/influxdata/telegraf/plugins/inputs"
-	"strconv"
 )
 
 type Teamspeak struct {
@@ -17,27 +18,6 @@ type Teamspeak struct {
 
 	client    *ts3.Client
 	connected bool
-}
-
-func (ts *Teamspeak) Description() string {
-	return "Reads metrics from a Teamspeak 3 Server via ServerQuery"
-}
-
-const sampleConfig = `
-  ## Server address for Teamspeak 3 ServerQuery
-  # server = "127.0.0.1:10011"
-  ## Username for ServerQuery
-  username = "serverqueryuser"
-  ## Password for ServerQuery
-  password = "secret"
-  ## Nickname of the ServerQuery client
-  nickname = "telegraf"
-  ## Array of virtual servers
-  # virtual_servers = [1]
-`
-
-func (ts *Teamspeak) SampleConfig() string {
-	return sampleConfig
 }
 
 func (ts *Teamspeak) connect() error {