+++ plugins/inputs/win_services/win_services.go
@@ -79,18 +79,6 @@ func (rmr *MgProvider) Connect() (WinServiceManager, error) {
 	}
 }
 
-var sampleConfig = `
-  ## Names of the services to monitor. Leave empty to monitor all the available services on the host. Globs accepted.
-  service_names = [
-    "LanmanServer",
-	"TermService",
-	"Win*",
-  ]
-  #excluded_service_names = [] # optional, list of service names to exclude
-`
-
-var description = "Input plugin to report Windows services info."
-
 //WinServices is an implementation if telegraf.Input interface, providing info about Windows Services
 type WinServices struct {
 	Log telegraf.Logger
@@ -117,14 +105,6 @@ func (m *WinServices) Init() error {
 	}
 
 	return nil
-}
-
-func (m *WinServices) Description() string {
-	return description
-}
-
-func (m *WinServices) SampleConfig() string {
-	return sampleConfig
 }
 
 func (m *WinServices) Gather(acc telegraf.Accumulator) error {