+++ plugins/outputs/amon/amon.go
@@ -22,17 +22,6 @@ type Amon struct {
 	client *http.Client
 }
 
-var sampleConfig = `
-  ## Amon Server Key
-  server_key = "my-server-key" # required.
-
-  ## Amon Instance URL
-  amon_instance = "https://youramoninstance" # required
-
-  ## Connection timeout.
-  # timeout = "5s"
-`
-
 type TimeSeries struct {
 	Series []*Metric `json:"series"`
 }
@@ -104,14 +93,6 @@ func (a *Amon) Write(metrics []telegraf.Metric) error {
 	}
 
 	return nil
-}
-
-func (a *Amon) SampleConfig() string {
-	return sampleConfig
-}
-
-func (a *Amon) Description() string {
-	return "Configuration for Amon Server to send metrics to."
 }
 
 func (a *Amon) authenticatedURL() string {